[android] #2726 - moved utility classes to separate package
