+++ android/MapboxGLAndroidSDKTestApp/src/main/java/com/mapbox/mapboxsdk/testapp/utils/TimingLogger.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mapbox.mapboxsdk.testapp;
+package com.mapbox.mapboxsdk.testapp.utils;
 
 import java.util.ArrayList;
